Update ActorActionFactory.cs
On line 275, endName is set to be empty. However, if animateAction.End is a value, then the endName should take that value to the end. The foreach loop (273 - 283) has the purpose of enabling the animation name to play to be set as a parameter value of a parameter in the story model. However, this is not the only way animateAction.End is used, so we must preserve endName if the condition at line 277 is never met. Thanks!